Reorganise and rename ``sphinx.ext.autodoc._importer``
