+++ sphinx/ext/autosummary/generate.py
@@ -33,9 +33,9 @@ from sphinx import __display_version__, package_dir
 from sphinx.builders import Builder
 from sphinx.config import Config
 from sphinx.errors import PycodeError
+from sphinx.ext.autodoc._importer import _import_module
 from sphinx.ext.autodoc._member_finder import _filter_enum_dict, unmangle
 from sphinx.ext.autodoc._sentinels import INSTANCE_ATTR, SLOTS_ATTR
-from sphinx.ext.autodoc.importer import import_module
 from sphinx.ext.autodoc.mock import ismock, undecorate
 from sphinx.ext.autosummary import (
     ImportExceptionGroup,
@@ -590,7 +590,7 @@ def _get_modules(
             continue
         fullname = f'{name}.{modname}'
         try:
-            module = import_module(fullname)
+            module = _import_module(fullname)
         except ImportError:
             pass
         else:
